requirements.txt
Flask~=3.1.1
Flask-SQLAlchemy
Flask-Migrate~=4.1.0
Flask-Admin~=1.6.1
Flask-SocketIO~=5.5.1
psycopg2-binary
redis~=6.2.0
google-api-python-client~=2.171.0
google-auth
python-dotenv~=1.1.0
eventlet

Werkzeug~=3.1.3
protobuf~=6.31.1